Use --docker-image instead of -o for clarity
We found in usability testing that people who were pushing a Docker image for
the first time were confused by the -o option in the examples. It took them a
while to work out what it meant.

First time users don't understand how the thing works and come with all sorts
of ideas about how it *might* work. For example. one participant wondered if CF
might accept a github repository name and look for a Dockerfile in it to build
an image. This is a reasonable hypothesis, consistent with previous
experience of CF from the buildpack lifecycle.

The -o option doesn't give them any clues that their hypothesis about how it
works is incorrect.

Change it to the long --docker-image argument to try to make it more explicit.